Move db fixtures into conftest
